Exercise2/Activity3/requirements.txt
psycopg[binary]
kafka-python